feat(dev-portal): move application list and details under dashboard route
