+++ gravitee-apim-portal-webui-next/src/app/dashboard/application-details/application-tab-logs/application-log-messages/application-log-messages.component.scss
