+++ gravitee-apim-portal-webui-next/src/app/dashboard/application-details/application-tab-logs/application-log-table/application-log-table.component.spec.ts
