+++ gravitee-apim-portal-webui-next/src/app/dashboard/application-details/application-tab-logs/application-tab-logs.service.ts
