+++ gravitee-apim-portal-webui-next/src/app/dashboard/application-details/application-tab-logs/message-log-detail-dialog/message-log-detail-dialog.component.ts
