+++ gravitee-apim-portal-webui-next/src/app/dashboard/application-details/application-tab-logs/more-filters-dialog/more-filters-dialog.component.scss
