+++ gravitee-apim-portal-webui-next/src/app/dashboard/application-details/application-tab-settings/application-tab-settings-read/application-tab-settings-read.component.scss
