+++ gravitee-apim-portal-webui-next/src/app/dashboard/application-details/application-tab-settings/application-tab-settings.component.html
