+++ gravitee-apim-portal-webui-next/src/app/dashboard/application-details/application.component.html
