+++ gravitee-apim-portal-webui-next/src/app/dashboard/applications/applications.component.html
