+++ gravitee-apim-portal-webui-next/src/app/dashboard/applications/applications.component.scss
@@ -13,44 +13,45 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-@use '../../scss/layout' as layout;
+@use '../../../scss/layout' as layout;
 
 :host {
   @include layout.large-container-legacy();
 }
 
-mat-card {
-  width: 100%;
-}
-
-.tile {
-  display: flex;
+.content {
   flex-flow: column;
-  gap: 16px;
+  gap: 32px;
 }
 
-.app-list-header {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  justify-content: space-between;
-
-  &__title {
+.apps-list {
+  &__header {
     display: flex;
-    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+    gap: 16px;
   }
 }
 
-.app-list__container {
-  display: grid;
-  gap: 32px;
-  grid-template-columns: repeat(3, minmax(0, 1fr));
+.list-page__pagination {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-top: 16px;
 
   &--mobile {
-    grid-template-columns: 1fr !important;
+    flex-direction: column;
+    gap: 16px;
   }
 
-  > * {
-    height: 256px;
+  &__page-size {
+    display: flex;
+    align-items: center;
+    gap: 8px;
+
+    mat-form-field {
+      width: 80px;
+    }
   }
 }
+
