+++ gravitee-apim-portal-webui-next/src/components/user-avatar/user-avatar.component.html
@@ -19,7 +19,7 @@
   {{ initials }}
 </button>
 <mat-menu #user="matMenu">
-  <button i18n="@@applications" mat-menu-item [routerLink]="['applications']">Applications</button>
+  <button i18n="@@applications" mat-menu-item [routerLink]="['dashboard', 'applications']">Applications</button>
   <button i18n="@@subscriptions" mat-menu-item [routerLink]="['dashboard', 'subscriptions']">Subscriptions</button>
   <button i18n="@@userAvatarMenuLogOut" mat-menu-item [routerLink]="['log-out']">Log out</button>
 </mat-menu>
